src/requirements.txt
python-can~=4.0.0